Updated SetupGraphics to set graphics to software
Will set graphics to Software mode, Min, and Fixed settings. Even if its
on OpenGL/DirectX mode and color finding won't work this will change it
to Soft. But you have to be at the login screen. Because we can't detect
if you are at the lobby and whether to use different coordinates